+++ R/response_subgroups.R
@@ -96,15 +96,6 @@ extract_rsp_subgroups <- function(variables,
                                   conf_level = 0.95,
                                   method = NULL,
                                   label_all = "All Patients") {
-  if ("strat" %in% names(variables)) {
-    warning(
-      "Warning: the `strat` element name of the `variables` list argument to `extract_rsp_subgroups() ",
-      "was deprecated in tern 0.9.3.\n  ",
-      "Please use the name `strata` instead of `strat` in the `variables` argument."
-    )
-    variables[["strata"]] <- variables[["strat"]]
-  }
-
   df_prop <- h_proportion_subgroups_df(
     variables,
     data,